lib/interface: Move interface to library
